// src/app/model/product.module.ts
export class ProductModule {
  constructor(number: number, apple: string, fruits: string, jpg: string, number2: number, healthyStuff: string, number3: number) {

  }

  id: number;
  name: string;
  category?: string;
  iamge?: any;
  price: number;
  description?: string;
  amount: number;
  quantity: number;
  //if i put ? means that is an options !
}